refactor: move `valid_text()` to `pdf_image_utils.py`
